Adding video cut and flask application

+++ youtube_video_handler.py
@@ -13,6 +13,15 @@ class YouTubeVideoHandler:
         self.__baseDirectory = baseDirectory
         self.__env = Env()
         self.__googleApi = GoogleApi()
+    
+    def getVideoTitle(self):
+        return self.__youtubeObject.title
+
+    def getVideoId(self):
+        return self.__youtubeObject.video_id
+    
+    def chanceBaseDir(self, dir):
+        self.__baseDirectory = dir
 
     def downloadAudio(self) -> AudioFileClip:
         audioFileName = os.path.join(
@@ -22,14 +31,14 @@ class YouTubeVideoHandler:
         audioFiles.get_by_itag(audio160Kbps).download(filename=audioFileName)
         return AudioFileClip(audioFileName)
 
-    def downloadVideo(self, insertAudio: bool) -> str:
+    def downloadVideo(self, insertAudio: bool, saveFile: bool = True) -> VideoFileClip:
         fullHdTag = 137
 
         videoFileName = os.path.join(
             self.__baseDirectory, f'{self.__youtubeObject.title}.mp4')
         finalVideoFileName = os.path.join(
             self.__baseDirectory, f'{self.__youtubeObject.title}_final.mp4')
-
+        
         self.__youtubeObject.streams.\
             filter(file_extension="mp4").\
             get_by_itag(fullHdTag).\
@@ -42,34 +51,30 @@ class YouTubeVideoHandler:
         
         if insertAudio:
             audioClip : AudioFileClip = self.downloadAudio()            
-            finalVideoClip = self.linkAudioToVideo(videoFileClip=videoClip, audioFileClip=audioClip, removeAudioFromFilesystem=True)        
+            finalVideoClip = self.linkAudioToVideo(videoFileClip=videoClip, audioFileClip=audioClip)        
         else: 
             finalVideoClip = videoClip
-        
-        finalVideoClip.write_videofile(finalVideoFileName)
+
+        if saveFile:
+            finalVideoClip.write_videofile(finalVideoFileName)
+
         if audioClip:
             self.removeAudioFile(audioClip)
-        # finalVideoClip: VideoFileClip = videoClip.set_audio(audioClip)
-         
-
-        # audioFilename = audioClip.filename
-        # audioClip.close()
-        # os.remove(audioFilename)
+                
         
-        
-        videoClip.close()
-        os.remove(videoFileName)
+        if saveFile:
+            videoClip.close()
+            os.remove(videoFileName)
+            
+        return finalVideoClip
     
     def removeAudioFile(self, audioFileClip: AudioFileClip):
         audioFileClip.close()
         os.remove(audioFileClip.filename)
 
 
-    def linkAudioToVideo(self, videoFileClip: VideoFileClip, audioFileClip: AudioFileClip, removeAudioFromFilesystem: bool) -> VideoFileClip:
+    def linkAudioToVideo(self, videoFileClip: VideoFileClip, audioFileClip: AudioFileClip) -> VideoFileClip:
         result = videoFileClip.set_audio(audioFileClip)        
-        # if removeAudioFromFilesystem:
-        #     audioFileClip.close()
-        #     os.remove(audioFileClip.filename)
         return result
 
 
@@ -88,9 +93,20 @@ class YouTubeVideoHandler:
                         
             with open(filePathSTR, 'w') as fileStr:
                 fileStr.writelines(strFormatter.format_transcript(caption))
-
+            
             return filePathSTR
         except Exception as e:
             logging.error(e)
             return ""
 
+    def cut(self, start: str, end: str, video: VideoFileClip, audio: AudioFileClip, fileName: str) -> None:
+        #00:00:09,640
+        hours, minutes, seconds = start.split(":")
+        startInSeconds = (int(hours) * 3600) + (int(minutes) * 60) + float(seconds.replace(",", "."))
+        hours, minutes, seconds = end.split(":")
+        endInSecodns = (int(hours) * 3600) + (int(minutes) * 60) + float(seconds.replace(",", "."))
+        videoClip: VideoFileClip = video.subclip(startInSeconds, endInSecodns)
+        audioClip: AudioFileClip = audio.subclip(start, end)
+        videoClipWithAudio: VideoFileClip = videoClip.set_audio(audioClip)
+        videoClipWithAudio.write_videofile(os.path.join(os.path.dirname(videoClip.filename), fileName)) 
+
